requirements.txt
# Core Dependencies for Resume Analyzer CLI
# PDF/DOCX Parsing
pdfplumber==0.10.3
python-docx==1.1.0
PyPDF2==3.0.1

# NLP & ML
nltk==3.8.1
scikit-learn==1.3.2
transformers==4.35.2
sentence-transformers==2.2.2
torch==2.1.1

# NER & Text Processing
spacy==3.7.2

# Embeddings & Similarity
numpy==1.26.2
scipy==1.11.4

# Report Generation
fpdf2==2.7.6
matplotlib==3.8.2
seaborn==0.13.0

# Data Handling
pandas==2.1.3
joblib==1.3.2

# OCR (Optional)
pytesseract==0.3.10
Pillow==10.1.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
python-dateutil==2.8.2
tqdm==4.66.1
colorama==0.4.6

# Streamlit Web App
streamlit==1.29.0
plotly==5.18.0
kaleido==0.2.1
